[Feat] #72 - Grouping MyBadge
