--- src/pages/dumpBadge.js
@@ -1,60 +1,0 @@
-const dumpBadge = [
-    {
-        id: 1,
-        title: 'Python 최고',
-        imageUrl: 'https://yhs.kr/static/image/python.svg',
-        date: new Date(),
-        active: false,
-        check: false,
-    },
-    {
-        id: 2,
-        title: 'Kotlin 최고',
-        imageUrl: 'https://yhs.kr/static/image/kotlin.svg',
-        date: new Date(),
-        active: true,
-        check: false,
-    },
-    {
-        id: 2,
-        title: 'Kotlin 최고',
-        imageUrl: 'https://yhs.kr/static/image/kotlin.svg',
-        date: new Date(),
-        active: false,
-        check: false,
-    },
-    {
-        id: 3,
-        title: 'Kotlin 최고',
-        imageUrl: 'https://yhs.kr/static/image/kotlin.svg',
-        date: new Date(),
-        active: true,
-        check: false,
-    },
-    {
-        id: 4,
-        title: 'Kotlin 최고',
-        imageUrl: 'https://yhs.kr/static/image/kotlin.svg',
-        date: new Date(),
-        active: true,
-        check: false,
-    },
-    {
-        id: 5,
-        title: 'Kotlin 최고',
-        imageUrl: 'https://yhs.kr/static/image/kotlin.svg',
-        date: new Date(),
-        active: true,
-        check: false,
-    },
-    {
-        id: 6,
-        title: 'Kotlin 최고',
-        imageUrl: 'https://yhs.kr/static/image/kotlin.svg',
-        date: new Date(),
-        active: true,
-        check: false,
-    },
-];
-
-export default dumpBadge;
