+++ src/pages/mybadge/dumpBadge.js
@@ -1,0 +1,11 @@
+const dumpBadge = [
+    {
+        title: 'Python 최고',
+        image: 'https://yhs.kr/static/image/python.svg',
+        date: new Date(),
+        active: false,
+        check: false,
+    },
+];
+
+export default dumpBadge;
